backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# MongoDB database
motor==3.3.2
pymongo==4.6.0

# AI and ML
google-generativeai==0.3.2
numpy==1.26.2
scikit-learn==1.3.2

# HTTP client and scraping
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
fake-useragent==1.4.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# Google Services
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.110.0

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Data processing
pandas==2.1.4

# CORS support
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.13.0
flake8==6.1.0

# Additional dependencies for async operations
typing-extensions==4.8.0
